make better base enemy + has hp_bar, rigibody movement  - dosent have pahtfind
