--- game/Assets/Scripts/Enemys/base_enemy/enemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4f87612e438aa044eb17cd3396cb5d68
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
