--- game/Assets/Scripts/Enemys/base_enemy/enemy_ai.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7a29ae27597502e4d8dff619a85717db
