+++ game/Assets/Scripts/Enemys/base_enemy/enemy_controller.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9017a1e225220724386734972eeec8e0
