--- game/Assets/Scripts/Player/combat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c31eb38c3fbc35644bc2776c71822f9b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
