--- game/Assets/Scripts/Player/combat/combat_controller.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a87fd02ad062d3347aa200705fc83abf
