--- game/Assets/Scripts/Player/hp.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cf5195d759016fe4b8a63bbbc1ad43d6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
