--- game/Assets/Scripts/Player/sound/audio_manager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fd2baeafa43e8f24494c849b135fd9aa
