--- game/Assets/Scripts/Player/sound/sound.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-[System.Serializable]
-public class sound
-{
-    public string name;
-    public AudioClip clip;
-}
