--- game/Assets/Scripts/Player/sound/sound.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6e505ed6a1b0b4743b96a2d5746ae6c9
