+++ game/Assets/Scripts/hp/hp_bar_ui.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fb968ce027b806b4096a19f2881fd4b2
