+++ game/Assets/Scripts/hp/hp_system.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6204f7f4e1ce66d4ebded0d89a76c406
