+++ game/Assets/Scripts/hp/ui_follow_player.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 13b19b7222a2dbf40af2a12f47491756
