+++ game/Assets/Scripts/managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f7ab29acb95cd504681cfb02842869c6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
