+++ game/Assets/Scripts/managers/player_manager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d802a6a57e0687645aed2a2d408edaf1
