+++ game/Assets/Scripts/sound.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da099ea8f884bfa4e85722eb0e95be77
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
