+++ game/Assets/Scripts/sound/sound.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+[System.Serializable]
+public class sound
+{
+    public string name;
+    public AudioClip clip;
+}
