+++ game/ProjectSettings/TagManager.asset
@@ -16,7 +16,7 @@ TagManager:
   - UI
   - fps_camera
   - enemy_1
-  - 
+  - wall
   - 
   - 
   -
